requirements.txt
# ═══════════════════════════════════════════════════════════════════════════════
# Foretyx Data Plane — Pinned Dependencies
# ═══════════════════════════════════════════════════════════════════════════════

# ── Core framework ────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.1

# ── PII detection (Microsoft Presidio) ────────────────────────────────────────
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
spacy==3.7.4

# ── ML injection guard (ONNX) ────────────────────────────────────────────────
onnxruntime==1.19.2
transformers==4.45.0

# ── HTTP client (async) ──────────────────────────────────────────────────────
httpx==0.27.2

# ── LLM provider ─────────────────────────────────────────────────────────────
google-generativeai>=0.8.0

# ── Configuration ────────────────────────────────────────────────────────────
python-dotenv==1.0.0

# ── Structured logging ───────────────────────────────────────────────────────
structlog>=24.0.0

# ── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
